requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

sqlmodel==0.0.14
asyncpg==0.29.0
aiosqlite==0.19.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

pydantic==2.5.0
pydantic-settings==2.1.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1
python-dotenv==1.0.0

pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
